requirements.txt
mistralai 
python-dotenv 
streamlit
